feat: wip removal of wyre
