--- src/components/add-cash/index.js
@@ -1,6 +1,0 @@
-export { default as AddCashFooter } from './AddCashFooter';
-export { default as AddCashForm } from './AddCashForm';
-export { default as AddCashStatus } from './AddCashStatus';
-export { default as ApplePayButton } from './ApplePayButton';
-export { default as NeedHelpButton } from './NeedHelpButton';
-export { default as SupportButton } from './SupportButton';
