+++ src/navigation/Routes.android.js
@@ -34,7 +34,6 @@ import WalletConnectRedirectSheet from '../screens/WalletConnectRedirectSheet';
 import WalletDiagnosticsSheet from '../screens/WalletDiagnosticsSheet';
 import WelcomeScreen from '../screens/WelcomeScreen';
 import WithdrawModal from '../screens/WithdrawModal';
-import WyreWebview from '../screens/WyreWebview';
 import RegisterENSNavigator from './RegisterENSNavigator';
 import { SwipeNavigator } from './SwipeNavigator';
 import { createBottomSheetNavigator } from './bottom-sheet';
@@ -45,7 +44,6 @@ import {
   restoreSheetConfig,
   stackNavigationConfig,
   learnWebViewScreenConfig,
-  wyreWebviewOptions,
 } from './config';
 import {
   addWalletNavigatorPreset,
@@ -80,21 +78,6 @@ const Stack = createStackNavigator();
 const OuterStack = createStackNavigator();
 const AuthStack = createStackNavigator();
 const BSStack = createBottomSheetNavigator();
-
-function AddCashFlowNavigator() {
-  const { colors } = useTheme();
-  const themedWyreWebviewOptions = useMemo(() => wyreWebviewOptions(colors), [
-    colors,
-  ]);
-  return (
-    <Stack.Navigator
-      initialRouteName={Routes.WYRE_WEBVIEW}
-      screenOptions={themedWyreWebviewOptions}
-    >
-      <Stack.Screen component={WyreWebview} name={Routes.WYRE_WEBVIEW} />
-    </Stack.Navigator>
-  );
-}
 
 function MainNavigator() {
   const initialRoute = useContext(InitialRouteContext);
@@ -189,10 +172,6 @@ function MainNavigator() {
         component={WelcomeScreen}
         name={Routes.WELCOME_SCREEN}
         options={{ animationEnabled: false, gestureEnabled: false }}
-      />
-      <Stack.Screen
-        component={AddCashFlowNavigator}
-        name={Routes.WYRE_WEBVIEW_NAVIGATOR}
       />
     </Stack.Navigator>
   );
