--- src/references/wyre/supportedCountries.js
@@ -1,118 +1,0 @@
-export const wyreSupportedCountries = {
-  AU: {
-    emojiName: 'flag_australia',
-    name: 'Australia',
-  },
-  AT: {
-    emojiName: 'flag_austria',
-    name: 'Austria',
-  },
-  BE: {
-    emojiName: 'flag_belgium',
-    name: 'Belgium',
-  },
-  CA: {
-    emojiName: 'flag_canada',
-    name: 'Canada',
-  },
-  CZ: {
-    emojiName: 'flag_czechia',
-    name: 'Czech Republic',
-  },
-  DK: {
-    emojiName: 'flag_denmark',
-    name: 'Denmark',
-  },
-  EE: {
-    emojiName: 'flag_estonia',
-    name: 'Estonia',
-  },
-  FI: {
-    emojiName: 'flag_finland',
-    name: 'Finland',
-  },
-  FR: {
-    emojiName: 'flag_france',
-    name: 'France',
-  },
-  DE: {
-    emojiName: 'flag_germany',
-    name: 'Germany',
-  },
-  GR: {
-    emojiName: 'flag_greece',
-    name: 'Greece',
-  },
-  HK: {
-    emojiName: 'flag_hong_kong_sar_china',
-    name: 'Hong Kong',
-  },
-  IE: {
-    emojiName: 'flag_ireland',
-    name: 'Ireland',
-  },
-  IT: {
-    emojiName: 'flag_italy',
-    name: 'Italy',
-  },
-  LV: {
-    emojiName: 'flag_latvia',
-    name: 'Latvia',
-  },
-  LT: {
-    emojiName: 'flag_lithuania',
-    name: 'Lithuania',
-  },
-  LU: {
-    emojiName: 'flag_luxembourg',
-    name: 'Luxembourg',
-  },
-  MX: {
-    emojiName: 'flag_mexico',
-    name: 'Mexico',
-  },
-  NL: {
-    emojiName: 'flag_netherlands',
-    name: 'The Netherlands',
-  },
-  NZ: {
-    emojiName: 'flag_new_zealand',
-    name: 'New Zealand',
-  },
-  NO: {
-    emojiName: 'flag_norway',
-    name: 'Norway',
-  },
-  PL: {
-    emojiName: 'flag_poland',
-    name: 'Poland',
-  },
-  PT: {
-    emojiName: 'flag_portugal',
-    name: 'Portugal',
-  },
-  SK: {
-    emojiName: 'flag_slovakia',
-    name: 'Slovakia',
-  },
-  SI: {
-    emojiName: 'flag_slovenia',
-    name: 'Slovenia',
-  },
-  ES: {
-    emojiName: 'flag_spain',
-    name: 'Spain',
-  },
-  SE: {
-    emojiName: 'flag_sweden',
-    name: 'Sweden',
-  },
-  CH: {
-    emojiName: 'flag_switzerland',
-    name: 'Switzerland',
-  },
-  US: {
-    emojiName: 'flag_united_states',
-    name: 'United States',
-  },
-};
